proper language list creation depends on the ipkg update

+++ trunk/package/webif/files/usr/lib/webif/webif-mklanglist.sh
@@ -1,12 +1,29 @@
 #!/bin/sh
+LANGLIST_NEEDS_UPDATE=0
+languages_root="/etc/languages.root"
 languages_lst="/etc/languages.lst"
-tmplanglst=$(mktemp "/tmp/.webif-XXXXXX")
-echo "option|en|English" >  "$tmplanglst"
-ipkg list | awk '/webif-lang/ { gsub("webif-lang-",""); print "option|"$1"|"$5}' | sort | uniq >> "$tmplanglst"
-if [ "'cat $tmplanglst'" != "option|en|English" ]; then
-	if [ "'cat $tmplanglst'" != "'cat $languages_lst'" ]; then
-		rm -f "$languages_lst"
-		mv -f "$tmplanglst" "$languages_lst"
-	fi
+if [ ! -f "$languages_lst" ]; then
+	cp -pf "$languages_root" "$languages_lst" 2>/dev/null
 fi
-rm -f "$tmplanglst"
+# rebuild the language list only after ipkg update
+for listfile in $(ls /usr/lib/ipkg/lists/* 2>/dev/null); do
+	[ "$listfile" -nt "$languages_lst" ] && LANGLIST_NEEDS_UPDATE=1
+done
+if [ "$LANGLIST_NEEDS_UPDATE" -eq "1" ]; then
+	tmplanglst=$(mktemp "/tmp/.webif-XXXXXX")
+	version_current=$(ipkg list_installed webif | grep "^webif -" | cut -d' ' -f3 | sed 's/\./\./')
+	for lngpkg in $(ipkg list webif-lang-* | sed "/webif-lang-[^[:space:]]* - ${version_current} -/!d; s/webif-lang-\([^[:space:]]*\)[[:space:]]-[[:space:]]${version_current}[[:space:]]-[[:space:]]\([^[:space:]]*\).*/option\|\1\|\2/"); do
+		if [ -n "$lngpkg" ]; then
+	                lngshort=$(echo "$lngpkg" | cut -d'|' -f1,2)
+        	        newlang=$(grep "^$lngshort" "$languages_root")
+	                if [ -z "$newlang" ]; then
+	                        echo "$lngpkg" >>"$tmplanglst"
+	                fi
+		fi
+	done
+	cat "$languages_root" | grep -v "^option|en|" >>"$tmplanglst"
+	rm -f "$languages_lst"
+	echo "option|en|English" >"$languages_lst"
+	cat "$tmplanglst" | sort | uniq | grep -v "^$" >>"$languages_lst"
+	rm -f "$tmplanglst"
+fi
